requirements.txt
annotated-types==0.7.0
argcomplete==3.5.3
certifi==2024.12.14
cfgv==3.4.0
charset-normalizer==3.4.1
click==8.1.8
click-option-group==0.5.6
colorama==0.4.6
commitizen==4.1.0
decli==0.6.2
distlib==0.3.9
dotty-dict==1.3.1
filelock==3.16.1
gitdb==4.0.12
GitPython==3.1.44
identify==2.6.5
idna==3.10
importlib_resources==6.5.2
Jinja2==3.1.5
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
nodeenv==1.9.1
packaging==24.2
platformdirs==4.3.6
pre_commit==4.0.1
prompt_toolkit==3.0.48
pydantic==2.8.2
pydantic_core==2.20.1
pygame==2.6.0
Pygments==2.19.1
python-dotenv==1.0.1
python-gitlab==4.13.0
python-semantic-release==9.15.2
PyTMX==3.32
PyYAML==6.0.2
questionary==2.1.0
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.4
shellingham==1.5.4
smmap==5.0.2
termcolor==2.5.0
tomlkit==0.13.2
typing_extensions==4.12.2
urllib3==2.3.0
virtualenv==20.28.1
wcwidth==0.2.13
Flask~=3.1.0
setuptools~=65.5.0
PyJWT~=2.10.1
pymongo~=3.11.0